Wire AZ-latency scenario to petsearch + dashboard AZ split
Make the az-app-slowdown FIS scenario observable end-to-end:
- petsearch now emits service/az/instance on every request log line via a
  shared RuntimeContext (read directly, not MDC, since the OTel Java agent
  doesn't preserve start-time MDC). AZ resolves from ECS Task Metadata v4.
- Latency-by-AZ dashboard: petsearch widget now splits p99 by AZ (parsed from
  the log), so the targeted AZ's latency spikes visibly during the experiment
  while the other AZ stays flat.
- az-app-slowdown template targets the workshop-standard AzImpairmentPower=Ready
  tag (already applied by CDK to the per-AZ PetSearch EC2 nodes) instead of a
  bespoke tag — so no extra tagging step is needed; setup script note updated.

petsearch is ECS-EC2 (one instance per AZ) by the workshop's existing design;
SSM network-latency only works on EC2, so it is the natural target for this
scenario. Validated live: petsearch logs split cleanly by AZ.
