+++ PetAdoptions/cdk/pet_stack/lib/dashboards/latency_az_dashboard.ts
@@ -49,8 +49,8 @@ export class LatencyByAZDashboard extends Construct {
                 logWidget('petlistadoptions — p99 latency by AZ',
                     "SOURCE '/ecs/PetListAdoptions' | fields az, latency_ms | filter ispresent(latency_ms) | stats pct(latency_ms,99) as p99_ms by az, bin(1m)",
                     12, 2, 12, 6),
-                logWidget('petsearch — p99 latency',
-                    "SOURCE '/ecs/PetSearchEc2' | parse @message /latency_ms=(?<lat>\\d+)/ | filter ispresent(lat) | stats pct(lat,99) as p99_ms, avg(lat) as avg_ms by bin(1m)",
+                logWidget('petsearch — p99 latency by AZ (FIS AZ-slowdown target)',
+                    "SOURCE '/ecs/PetSearchEc2' | parse @message /az=(?<az>[a-z0-9-]*)/ | parse @message /latency_ms=(?<lat>\\d+)/ | filter ispresent(lat) | stats pct(lat,99) as p99_ms, avg(lat) as avg_ms by az, bin(1m)",
                     0, 8, 12, 6),
                 logWidget('petsite — p99 latency by node (EKS)',
                     "SOURCE '/aws/containerinsights/PetSite/application' | filter ispresent(log_processed.State.LatencyMs) | stats pct(log_processed.State.LatencyMs,99) as p99_ms by log_processed.State.Node, bin(1m)",
